Reformatted all code.
This means that all tabs are now spaces, and curly braces may be in
different spots than you are used to. Also, the diff will show that every
single line that was affected by the reformat was deleted. This commit has
the exact same code as the previous, and you should refer to that for code
differences.